[dnn-library] FFT: Fixed a bug in twiddle_vector_small
The precalculation of the twiddle vector had a cast from size_t to
float, which is not supported by our compiler and silently generates
unsupported instructions.

+++ dnn-library/include/internal/FFT.h
@@ -137,7 +137,7 @@ INLINE_ATTR void w(size_t j, float recN, float& real, float& img) {
 
 INLINE_ATTR void twiddle_vector_small(size_t size, float real[], float img[]) {
   float recN = rec(size);
-  for (size_t i = 0; i < size; ++i) {
+  for (uint32_t i = 0; i < static_cast<uint32_t>(size); ++i) {
     w(i, recN, real[i], img[i]);
   }
 }
